src/requirements.txt
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-groq
chromadb
pypdf
python-dotenv
streamlit
pydantic
sentence-transformers